pyproject.toml
[project]
name = "dash-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=2.18.0",
    "dash-bootstrap-components>=2.0.4",
    "gunicorn>=23.0.0",
    "hdbscan>=0.8.41",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.48",
    "umap-learn>=0.5.11",
]
